Tools for the genetic code

aminoacid() returns the aminoacid generated by the informed codon(s).

gen_code() returns all the amino acids considering the translation table used.

aminoacid(codon, transl_table = 1, type = "olc")

gen_code(transl_table = 1)

Arguments

codon
A character vector (upper- or lower-case letters) with the 5'-3' codon (e.g., "AUG").

transl_table
The translation table used. Defaults to 1 (standard genetic code). See more in Details section.

type
The type of output. One of

"olc" for one-letter code, e.g., "V".

"tlc" for three-letter code, e.g., "Val".

"name" for the complete name, e.g., "Valine".

Value

aminoacid() returns a character vector in the format code_letter_name of the given aminoacid(s)

gen_code() returns a data.frame object with the following columns.

fb The first base of the codon.

sb The second base of the codon.

tb The third base of the codon.

codon The codon (5'-3').

olc The one letter code.

tlc The three letter code.

name The name of the aminoacid.

The stop codons are depicted by "*"

Details

The translation is performed according to the desired translation table. See more at https://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=cgencodes

The Standard Code (default)

The Vertebrate Mitochondrial Code

The Yeast Mitochondrial Code

The Mold, Protozoan, and Coelenterate Mitochondrial Code and the Mycoplasma/Spiroplasma Code

The Invertebrate Mitochondrial Code

The Ciliate, Dasycladacean and Hexamita Nuclear Code

The Echinoderm and Flatworm Mitochondrial Code

The Euplotid Nuclear Code

The Bacterial, Archaeal and Plant Plastid Code

The Alternative Yeast Nuclear Code

The Ascidian Mitochondrial Code

The Alternative Flatworm Mitochondrial Code

Chlorophycean Mitochondrial Code

Trematode Mitochondrial Code

Scenedesmus obliquus Mitochondrial Code

Thraustochytrium Mitochondrial Code

Rhabdopleuridae Mitochondrial Code

Candidate Division SR1 and Gracilibacteria Code

Pachysolen tannophilus Nuclear Code

Karyorelict Nuclear Code

Condylostoma Nuclear Code

Mesodinium Nuclear Code

Peritrich Nuclear Code

Blastocrithidia Nuclear Code

Cephalodiscidae Mitochondrial UAA-Tyr Code

Examples

aac <- c("AUG", "CGA", "UGA")
aminoacid(aac)
#> [1] "M" "R" "*"
aminoacid(aac, transl_table = 2)
#> [1] "M" "R" "W"

# complete genetic code
gen_code()
#>    fb sb tb codon olc tlc          name
#> 1   A  A  A   AAA   K Lys        Lysine
#> 2   A  A  U   AAU   N Asn    Asparagine
#> 3   A  A  G   AAG   K Lys        Lysine
#> 4   A  A  C   AAC   N Asn    Asparagine
#> 5   A  U  A   AUA   I Ile    Isoleucine
#> 6   A  U  U   AUU   I Ile    Isoleucine
#> 7   A  U  G   AUG   M Met    Methionine
#> 8   A  U  C   AUC   I Ile    Isoleucine
#> 9   A  G  A   AGA   R Arg      Arginine
#> 10  A  G  U   AGU   S Ser        Serine
#> 11  A  G  G   AGG   R Arg      Arginine
#> 12  A  G  C   AGC   S Ser        Serine
#> 13  A  C  A   ACA   T Thr     Threonine
#> 14  A  C  U   ACU   T Thr     Threonine
#> 15  A  C  G   ACG   T Thr     Threonine
#> 16  A  C  C   ACC   T Thr     Threonine
#> 17  U  A  A   UAA   *   *             *
#> 18  U  A  U   UAU   Y Tyr     Tyrosisne
#> 19  U  A  G   UAG   *   *             *
#> 20  U  A  C   UAC   Y Tyr     Tyrosisne
#> 21  U  U  A   UUA   L Leu       Leucine
#> 22  U  U  U   UUU   F Phe Phenylalanine
#> 23  U  U  G   UUG   L Leu       Leucine
#> 24  U  U  C   UUC   F Phe Phenylalanine
#> 25  U  G  A   UGA   *   *             *
#> 26  U  G  U   UGU   C Cys      Cysteine
#> 27  U  G  G   UGG   W Trp   Tryptophane
#> 28  U  G  C   UGC   C Cys      Cysteine
#> 29  U  C  A   UCA   S Ser        Serine
#> 30  U  C  U   UCU   S Ser        Serine
#> 31  U  C  G   UCG   S Ser        Serine
#> 32  U  C  C   UCC   S Ser        Serine
#> 33  G  A  A   GAA   E Glu  GlutamicAcid
#> 34  G  A  U   GAU   D Asp  AsparticAcid
#> 35  G  A  G   GAG   E Glu  GlutamicAcid
#> 36  G  A  C   GAC   D Asp  AsparticAcid
#> 37  G  U  A   GUA   V Val        Valine
#> 38  G  U  U   GUU   V Val        Valine
#> 39  G  U  G   GUG   V Val        Valine
#> 40  G  U  C   GUC   V Val        Valine
#> 41  G  G  A   GGA   G Gly       Glycine
#> 42  G  G  U   GGU   G Gly       Glycine
#> 43  G  G  G   GGG   G Gly       Glycine
#> 44  G  G  C   GGC   G Gly       Glycine
#> 45  G  C  A   GCA   A Ala       Alanine
#> 46  G  C  U   GCU   A Ala       Alanine
#> 47  G  C  G   GCG   A Ala       Alanine
#> 48  G  C  C   GCC   A Ala       Alanine
#> 49  C  A  A   CAA   Q Gln     Glutamine
#> 50  C  A  U   CAU   H His     Histidine
#> 51  C  A  G   CAG   Q Gln     Glutamine
#> 52  C  A  C   CAC   H His     Histidine
#> 53  C  U  A   CUA   L Leu       Leucine
#> 54  C  U  U   CUU   L Leu       Leucine
#> 55  C  U  G   CUG   L Leu       Leucine
#> 56  C  U  C   CUC   L Leu       Leucine
#> 57  C  G  A   CGA   R Arg      Arginine
#> 58  C  G  U   CGU   R Arg      Arginine
#> 59  C  G  G   CGG   R Arg      Arginine
#> 60  C  G  C   CGC   R Arg      Arginine
#> 61  C  C  A   CCA   P Pro       Proline
#> 62  C  C  U   CCU   P Pro       Proline
#> 63  C  C  G   CCG   P Pro       Proline
#> 64  C  C  C   CCC   P Pro       Proline