Translate a RNA sequence to a protein sequence

Translate a nucleotide (RNA) sequence to a protein chain according to the standard genetic code

translate(
  mrna,
  transl_table = 1,
  start_codon = "AUG",
  start_codon_first = TRUE,
  no_stop = FALSE,
  type = "olc",
  sep = ""
)

Arguments

mrna
The messenger RNA.

transl_table
The translation table used. Defaults to 1 (standard genetic code). See more at aminoacid()

start_codon
The start codon. Defaults to "AUG".

start_codon_first
Start the translation with the start_codon? defaults to TRUE. In this case, all bases before the start_codon will be ignored.

no_stop
A logical indicating if the translation shouldn't stop when a stop codon is encountered. Defaults to FALSE. If TRUE, translation will continue up to the end of the chain, and stop codons will be represented by "*".

type
The type of output. One of

"olc" for one-letter code, e.g., "V" (default).

"tlc" for three-letter code, e.g., "Val".

"name" for the complete name, e.g., "Valine".

sep
The separator between amino acids. Defaults to "-".

Value

A character

Examples

code <- rand_nuc_seq(25, type = "rna")
#> Start codon  : 1-3
#> Stop codon   : UAG
#> Position stop: 22-24
translate(code)
#> [1] "MLCHASK*"